requirements.txt
requests
python-dotenv
datadog_api_client